(PC-37953)[API] refactor: remove `cgr` connector
