+++ api/src/pcapi/core/external_bookings/cgr/serializers.py
